scale number of errors to resource count value (when all restarts exhausted).
Note that this will be cleared when new run is started even if all resources are still kept by the previous run (including cases where processes getting stuck and/or having prolonged crashes).
Delta of active_resources and active_resources_activeRun could be used to detect this case, as this is already available in resource_summary as well as archived in elasticsearch.

version 2.5.5 of fffmeta and hltd