protect against exception thrown when trying to unlock lock which is not locked.
It is not understood how this problem appears since acquire is always executed earlier and unconditionally.
It points to lock not being taken (silent failure?), but according to python documentation function should block if unable to take lock. Error
also appeared on multiple machines so it doesn't seem to be related to hardware failure.

Another possibility is that something happens with the file descriptor
underneath?

A log printing acquire return will be printed in such case in the future
(work is not yet finished), but we add this workaround in the meantime
in place where this was seen to happen.